few changes:
cron from 5min to 30min
sort group by most recent
save data split to functions

+++ save-data.js
@@ -4,23 +4,36 @@ const TelegramBot = require('node-telegram-bot-api');
 
 const bot = new TelegramBot(data.telegramToken, {polling: true});
 
-module.exports = function saveToFile(postData) {
+function isAlreadySaved(allData, postData) {
+    for(let i = 0; i < allData.postsData.length; i++)
+        if(allData.postsData[i].postUrl === postData.postUrl)
+            return true;
+
+    return false;
+}
+
+function isMatch(postData) {
+    return true;
+}
+
+function saveToFile(postData) {
     let allData = { postsData: [] };
-    let isAlreadySaved = false;
 
     if(fs.existsSync('postsData.json')) {
         allData = fs.readFileSync('postsData.json', 'utf8');
         allData = JSON.parse(allData);
     }
 
-    for(let i = 0; i < allData.postsData.length && !isAlreadySaved; i++) {
-        isAlreadySaved = (allData.postsData[i].postUrl === postData.postUrl);
-    }
-
-    if(!isAlreadySaved) {
+    if(!isAlreadySaved(allData, postData) && isMatch(postData)) {
         bot.sendMessage(data.channelId, postData.postUrl, {disable_web_page_preview: true});
         allData.postsData.push(postData);
         allData = JSON.stringify(allData);
         fs.writeFileSync('postsData.json', allData);
     }
+}
+
+module.exports = {
+    isAlreadySaved,
+    isMatch,
+    saveToFile
 }
